cleans up namespaces
