--- Source/Relativity.Test.Helpers/Relativity.Test.Helpers/Application/ApplicationInstallException.cs
@@ -1,9 +1,0 @@
-﻿using System;
-
-namespace Relativity.Test.Helpers.Application
-{
-	public class ApplicationInstallException : Exception
-	{
-		public ApplicationInstallException(string message) : base(message) { }
-	}
-}
