--- Source/Relativity.Test.Helpers/Relativity.Test.Helpers/ArtifactHelpers/Folders.cs
@@ -1,53 +1,0 @@
-﻿using kCura.Relativity.Client;
-using kCura.Relativity.Client.DTOs;
-using Relativity.API;
-using Relativity.Test.Helpers.ServiceFactory.Extentions;
-using System;
-using System.Linq;
-//using IServicesMgr = Relativity.Test.Helpers.Interface.IServicesMgr;
-using IServicesMgr = Relativity.API.IServicesMgr;
-
-namespace Relativity.Test.Helpers.ArtifactHelpers
-{
-
-    /// <summary>
-    /// 
-    /// Helpers to interact with Folders in Relativity
-    /// 
-    /// </summary>
-    /// 
-    public class Folders
-	{
-		public static Int32 GetRootFolderArtifactID(Int32 workspaceID, IServicesMgr svgMgr, string userName, string password)
-		{
-			using (IRSAPIClient client = svgMgr.GetProxy<IRSAPIClient>(new Models.ConfigurationModel()))
-			{
-				Query<Folder> query = new Query<Folder>();
-				query.Condition = new TextCondition(FolderFieldNames.Name, TextConditionEnum.EqualTo, WorkspaceHelpers.WorkspaceHelpers.GetWorkspaceName(client, workspaceID));
-				query.Fields = FieldValue.NoFields;
-				var ResultSet = client.Repositories.Folder.Query(query);
-
-				if (!ResultSet.Success)
-				{
-					throw new System.Exception("Folder was not found");
-				}
-				else if (ResultSet.TotalCount == 0)
-				{
-					throw new System.Exception("folder count was 0, so the folder was not found");
-				}
-				return ResultSet.Results.FirstOrDefault().Artifact.ArtifactID;
-			}
-		}
-
-
-		public static String GetFolderName(Int32 folderArtifactID, IDBContext workspaceDbContext)
-		{
-			string sql = String.Format("select Name from folder where ArtifactID = {0}", folderArtifactID);
-
-			string folderName = workspaceDbContext.ExecuteSqlStatementAsScalar(sql).ToString();
-
-			return folderName;
-		}
-
-	}
-}
