+++ Source/Relativity.Test.Helpers/Relativity.Test.Helpers/Exceptions/IntegrationTestException.cs
@@ -1,14 +1,7 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Relativity.Test.Helpers.Exceptions
+﻿namespace Relativity.Test.Helpers.Exceptions
 {
 	public class IntegrationTestException : System.Exception
 	{
-
 		public IntegrationTestException()
 		{
 		}
@@ -24,7 +17,7 @@ namespace Relativity.Test.Helpers.Exceptions
 		}
 
 		// A constructor is needed for serialization when an
-		// exception propagates from a remoting server to the client. 
+		// exception propagates from a remoting server to the client.
 		protected IntegrationTestException(System.Runtime.Serialization.SerializationInfo info, System.Runtime.Serialization.StreamingContext context)
 		{
 		}
