+++ Source/Relativity.Test.Helpers/Relativity.Test.Helpers/Import/Constants.cs
@@ -1,0 +1,8 @@
+﻿namespace Relativity.Test.Helpers.Import
+{
+    internal static class Constants
+    {
+        public const string NATIVE_FILE = "Native File";
+        public const string ENCODING_ISO_8859_1_NAME = "iso-8859-1";
+    }
+}
