+++ Source/Relativity.Test.Helpers/Relativity.Test.Helpers/Import/ImportDocument.cs
@@ -1,0 +1,8 @@
+﻿namespace Relativity.Test.Helpers.Import.Models
+{
+	public class ImportDocument
+	{
+		public string Name { get; set; }
+		public string FilePath { get; set; }
+	}
+}
