--- Source/Relativity.Test.Helpers/Relativity.Test.Helpers/ImportAPIHelper/Constants.cs
@@ -1,8 +1,0 @@
-﻿namespace Relativity.Test.Helpers.ImportAPIHelper
-{
-    internal static class Constants
-    {
-        public const string NATIVE_FILE = "Native File";
-        public const string ENCODING_ISO_8859_1_NAME = "iso-8859-1";
-    }
-}
