--- Source/Relativity.Test.Helpers/Relativity.Test.Helpers/ImportAPIHelper/ImportDocument.cs
@@ -1,8 +1,0 @@
-﻿namespace Relativity.Test.Helpers.ImportAPIHelper
-{
-	public class ImportDocument
-	{
-		public string Name { get; set; }
-		public string FilePath { get; set; }
-	}
-}
