+++ Source/Relativity.Test.Helpers/Relativity.Test.Helpers/Objects/Application/ApplicationInstallException.cs
@@ -1,0 +1,9 @@
+﻿using System;
+
+namespace Relativity.Test.Helpers.Application.Exceptions
+{
+	public class ApplicationInstallException : Exception
+	{
+		public ApplicationInstallException(string message) : base(message) { }
+	}
+}
