+++ Source/Relativity.Test.Helpers/Relativity.Test.Helpers/Objects/User/create.cs
@@ -1,0 +1,12 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Relativity.Test.Helpers.UserHelpers
+{
+	public class Create
+	{
+	}
+}
