+++ Source/Relativity.Test.Helpers/Relativity.Test.Helpers/Objects/Workspace/WorkspaceHelpers.cs
@@ -1,0 +1,24 @@
+﻿using kCura.Relativity.Client;
+using kCura.Relativity.Client.DTOs;
+
+namespace Relativity.Test.Helpers.Workspace
+{
+	public class WorkspaceHelpers
+	{
+		public static string GetWorkspaceName(IRSAPIClient client, int workspaceArtifactId)
+		{
+			var oldId = client.APIOptions.WorkspaceID;
+			try
+			{
+				client.APIOptions.WorkspaceID = -1;
+				kCura.Relativity.Client.DTOs.Workspace workspace = client.Repositories.Workspace.ReadSingle(workspaceArtifactId);
+				return workspace.Name;
+			}
+			finally
+			{
+				client.APIOptions.WorkspaceID = oldId;
+			}
+
+		}
+	}
+}
