+++ Source/Relativity.Test.Helpers/Relativity.Test.Helpers/ServiceFactory/Extentions/ServiceManagerExtension.cs
@@ -1,6 +1,6 @@
 ﻿using Relativity.API;
 using Relativity.Services.ServiceProxy;
-using Relativity.Test.Helpers.Models;
+using Relativity.Test.Helpers.Configuration.Models;
 using System;
 
 namespace Relativity.Test.Helpers.ServiceFactory.Extentions
