--- Source/Relativity.Test.Helpers/Relativity.Test.Helpers/UserHelpers/create.cs
@@ -1,12 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Relativity.Test.Helpers.UserHelpers
-{
-	public class Create
-	{
-	}
-}
